refactor: rearrange files of 'games' feature
