+++ src/features/games/GamesPage.tsx
@@ -1,8 +1,8 @@
 import { Grid, WithStyles, withStyles } from '@material-ui/core';
 import React from 'react';
 import { IGame } from 'types';
-import GameCard from './GameCard';
-import gamesPageStyles from './games-page-styles';
+import GameCard from './game-card/GameCard';
+import gamesPageStyles from './styles';
 
 interface Props extends WithStyles<typeof gamesPageStyles> {
   games: Array<IGame>;
